backend/notification_service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
motor==3.1.1
pymongo==4.3.3
python-dotenv==1.0.0
aiohttp==3.8.4
pydantic==1.10.7
python-multipart==0.0.6
email-validator==2.0.0 